Switch to Typescript
