+++ package.json
@@ -18,24 +18,30 @@
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^29.5.3",
+    "@types/node": "^20.4.5",
+    "@types/react": "^18.2.17",
+    "@types/react-dom": "^18.2.7",
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.8.1",
     "canvg": "^4.0.0",
     "fast-xml-parser": "^4.0.7",
     "file-saver": "^2.0.5",
     "jszip": "^3.7.1",
+    "minimist": ">=1.2.6",
+    "nth-check": ">=2.1.1",
+    "pcb-stackup": "^4.2.8",
     "react": "^17.0.2",
     "react-bootstrap": "^2.1.2",
     "react-dom": "^17.0.2",
     "react-scripts": "5.0.1",
     "title-case": "^3.0.3",
-    "nth-check": ">=2.1.1",
-    "minimist": ">=1.2.6"
+    "typescript": "^4.9.5"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
     "eject": "react-scripts eject",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build"
@@ -59,6 +65,9 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^4.0.0"
+    "buffer": "^6.0.3",
+    "gh-pages": "^4.0.0",
+    "process": "^0.11.10",
+    "react-app-rewired": "^2.2.1"
   }
 }
