+++ public/index.html
@@ -10,8 +10,6 @@
     <meta name="msapplication-TileColor" content="#da532c">
     <meta name="theme-color" content="#ffffff">
 
-    <script src="%PUBLIC_URL%/gerber-to-svg.min.js"></script>
-    <script src="%PUBLIC_URL%/pcb-stackup.min.js"></script>
     <title>Photonic Etcher</title>
   </head>
   <body>
