+++ src/renderer/stackup-renderer.ts
@@ -1,9 +1,25 @@
+import pcbStackup, { Options } from "pcb-stackup";
 import {SVGToImage} from "./svg_to_png";
+import { GerberFile } from "../ui/utils/pcb_api_utils";
+import gerberToSvg from "gerber-to-svg";
+import whatsThatGerber from "whats-that-gerber";
 
-const pcbStackup = window.pcbStackup;
-const renderConverter = window.gerberToSvg.render;
 
-export default async function renderStackup(layers, options){
+export interface StackupLayer{
+    id: number,
+    side: whatsThatGerber.GerberSide,
+    type: whatsThatGerber.GerberType,
+    filename?: string,
+    svg: string
+}
+export type Stackup ={combinedIMGs:{
+    top_svg: any,
+    bottom_svg: any,
+    top_png_blob_url: any,
+    bottom_png_blob_url: any
+},individualLayerSVGs:StackupLayer[]}
+
+export default async function renderStackup(layers: GerberFile[], options: Options): Promise<Stackup>{
     const stackup = await pcbStackup(layers, options);
     let id_num = 14214;
 
@@ -15,13 +31,13 @@ export default async function renderStackup(layers, options){
             layer.converter.width = stackup.top.width;
             layer.converter.height = stackup.top.height;
             id_num += 1;
-            const layerSVG = renderConverter(layer.converter, id_num.toString().padStart(12, '0'));
+            const svgString= gerberToSvg.render(layer.converter, id_num.toString().padStart(12, '0'));
             return {
                 "id": layer.id,
                 "side": layer.side,
                 "type": layer.type,
                 "filename": layer.filename,
-                "svg": layerSVG
+                "svg":svgString
         }
     });
 
@@ -32,5 +48,5 @@ export default async function renderStackup(layers, options){
         "bottom_png_blob_url": await SVGToImage({svg: stackup.bottom.svg, width: 1024})
     };
 
-    return [combinedIMGs, individualLayerSVGs];
+    return {combinedIMGs, individualLayerSVGs};
 }
