+++ src/ui/utils/util.ts
@@ -1,8 +1,8 @@
 import {useEffect, useRef} from "react";
 
 
-export function usePrevious(value) {
-    const ref = useRef();
+export function usePrevious<T>(value: T): T|undefined {
+    const ref = useRef<T>();
     useEffect(() => {
         ref.current = value;
     });
